Create README.md
This is a project I created originally based off the eTicket tutorial done by eTrupja.  I initially followed his youtube guide to create the eTicket ASP.NET application and then used that as a basis for the Video Game Store. I added additional functionality as I went and will continue to improve it as I use it as a practice application to improve my ASP.NET MVC skills.